GCP_Scripts/GCS_to_SQL/requirements.txt
functions-framework==3.*
google-cloud-storage
pandas
cloud-sql-python-connector
pymysql
sqlalchemy
numpy